Add name and nregs to functions
Add the name and nregs to the FunctionSpec. The name is just a string
used to lookup main entry points. The nregs is the number of registers,
used to allocate temporary slots for function calls.

Note that currently, the number of registers is not used, the VM always
allocates 10 slots for functions to use.

Add the name and nregs to the DlFunctionEntry structure, and update the
js compiler and DlLoader to use these fields properly.

Signed-off-by: Robert Young <rwy0717@gmail.com>