Properly calculate the register count of a function
The registers are indexable operand stack slots int the VM. Right now,
the registers correspond exactly to the local variables of a function.

This patch updates the function handler to store the number of registers
required into the function declaration. This happens at the end of the
function handler. Then, we defers the generation of the function table
entries later, after the functions have been handled.

Signed-off-by: Robert Young <rwy0717@gmail.com>