Rework NO_MORE_BYTECODES
Previously, NO_MORE_BYTECODES was the integer literal 0, and
Instructions had to be comparable against this special value. To
allow this, Instruction could degrade to uint32_t implicitly, which was
not nice.

Now, I've rename NO_MORE_BYTECODES to END_SECTION, a special constant
instance of an Instruction. This instruction is:

  {ByteCode::END_SECTION, 0}

Which translates to the RawInstruction constant zero.

I've added and documented a new END_SECTION bytecode, which is only used
here.

I've also updated the Instruction class with equality operators, to
allow for the comparison against the END_SECTION instruction, and
removed the operator RawInstruction code.

Signed-off-by: Robert Young <rwy0717@gmail.com>