Add a Cons type for building linked lists
Cons<T> is a linked-list element. Each cons is a pair, holding one
instance of T, and a pointer to the next Cons<T>. The Cons cells do not
manage their own memory. It is up to the user to do allocation and
freeing.

The RootRef object is now a wrapper around pushing and popping
Cons<Cell*> off a list.

Introducing the cons cell base class lets us write a general iteration
API, which should make it easier to roll new kinds of linked lists in
the future.

Signed-off-by: Robert Young <rwy0717@gmail.com>