Add CoreTypes to the SlotType structure
We're adding in a notion of core, or fundamental types.
CoreTypes allow us to strongly-type object slots, and lays the
groundwork for variable-width slots, and compressed pointers.

There are a collection of binary/data types, but of particular interest
are REF type, forGC-pointer slots, and VALUE, for untyped / self
encoded values.

Signed-off-by: Robert Young <rwy0717@gmail.com>