Update README [skip-ci]
This patch updates the README file to contain more up-to-date
information on building b9. There were some missing steps
that have now been addressed.

The README file contains instructions on how to build b9
on Ubuntu 16.04. The steps here make users update their
nodejs to the latest version to ensure all features being
used in the project are available.

Also, I switched the instructions to build out of source tree
as it is a cleaner approach in my opinion.

Signed-off-by: Nazim Uddin Bhuiyan <nazimudd@ualberta.ca>