Augment AST before generating module
Some of the AST nodes now have a functionEntry member that
provides a way to lookup symbols defined in parent scopes.
FunctionContext is used for context tracking as well
as looking up symbols beyond local context through the use
of a stack. Function indexing starts from 1, as index 0 is
reserved for the top-level body, which is now treated as a
function.

Also-by: Robert Young <rwy0717@gmail.com>
Signed-off-by: Nazim Uddin Bhuiyan <nazimudd@ualberta.ca>