Only call factorial up to ten
Previously, we were calling factorial on 1 to 20. factorial(20) results
in an integer that is too large to be represented in 48bits.

To support factorial(20), we have to implement promotion of integers to
doubles in the VM. We do not yet support doubles at all.

This calculation was underflowing in the interpreter, and causing badly
tagged values in the JIT.

Fixes the factorial test in CI.

Signed-off-by: Robert Young <rwy0717@gmail.com>